feat: 主世界全立方体方块 258 种（约原版 1/3）+ 矿石生成 + 创造选块界面
- blocks.ts 重写：tile 按名注册（interning），石头/深板岩/矿石/16 色羊毛
  混凝土陶瓦玻璃/8 种木材/海洋冰雪/功能块共 258 种（256 固体 + 空气/水），
  挖掘属性改为 BlockDef 驱动（tool/needsPick/unbreakable），基岩与强化
  深板岩生存不可破坏
- 矿石生成（lib/oregen.ts）：基岩层 + 深板岩渐变 + 煤/铁/铜/金/红石/青金石/
  钻石/绿宝石矿脉与岩石团簇，按 chunk 确定性；虚空地形跳过（测试安全）
- chunk 数据 Uint8 → Uint16（存档版本 4→5，旧档按既有策略清空）
- 创造模式：热键栏 store 化 + E 键选块界面（搜索/分类/256 种），放置预览
  与生存手持联动，ghost mesh 按 id 命令式创建（WebGPU 安全）
- 配方：每木种木板、石砖链、磨制岩石、深板岩链、砂岩、泥砖；熔炉烧炼
  （平滑石/裂纹砖/陶瓦/玻璃红沙/木炭全木种），燃料扩展（干海带块/煤块）
- 贴图：scripts/build-pack.ts 按 TILE_STEMS 从 Faithful 提取 286 格进
  pack/（平滑砂岩/木头/雪块等按原版规则取共享贴图），atlas 8×52
- 新增 blocks/oregen 单测 10 个（注册表完整性/确定性/虚空安全/矿脉宿主）

+++ lib/__tests__/structures.test.ts
@@ -42,7 +42,7 @@ describe('村庄结构', () => {
     const hut = villageStructures(sh, rx, rz, v.x, v.z).find((s) => s.type === 'hut')!;
     const cx = Math.floor(hut.x / CHUNK_SIZE);
     const cz = Math.floor(hut.z / CHUNK_SIZE);
-    const data = new Uint8Array(CHUNK_SIZE * CHUNK_SIZE * WORLD_HEIGHT);
+    const data = new Uint16Array(CHUNK_SIZE * CHUNK_SIZE * WORLD_HEIGHT);
     // 铺一块平地当地基
     const h = t.heightAt(hut.x, hut.z);
     for (let x = 0; x < CHUNK_SIZE; x++) {
